+++ src/test/java/org/opentmf/catalog/sync/CatalogSyncAutoConfigurationTests.java
@@ -1,11 +1,11 @@
 package org.opentmf.catalog.sync;
 
-import org.opentmf.catalog.sync.config.CatalogSyncAutoConfiguration;
-import org.opentmf.catalog.sync.config.CatalogSyncProperties;
-import org.opentmf.db.lock.service.api.DbLockService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
+import org.opentmf.catalog.sync.config.CatalogSyncAutoConfiguration;
+import org.opentmf.catalog.sync.config.CatalogSyncProperties;
+import org.opentmf.db.lock.service.api.DbLockService;
 import org.springframework.context.ApplicationContext;
 
 /**
@@ -15,7 +15,6 @@ class CatalogSyncAutoConfigurationTests {
 
   @Test
   void testCatalogSyncAutoConfiguration_withInvalidData_throwsException() {
-    // for getting rid of unused class.
     Assertions.assertThrows(Exception.class, () ->
         new CatalogSyncAutoConfiguration(
             Mockito.mock(ApplicationContext.class),